Prep SQLEditor to be moved out to experimental package
