+++ src/components/sql-editor/index.ts
@@ -1,0 +1,17 @@
+export { SQLEditor } from './components/SQLEditor';
+export { SQLEditorTestUtils } from './test-utils';
+export { LinkedToken } from './utils/LinkedToken';
+
+export {
+  TableDefinition,
+  ColumnDefinition,
+  StatementPlacementProvider,
+  SuggestionKindProvider,
+  LanguageCompletionProvider,
+  OperatorType,
+  TokenType,
+  StatementPosition,
+  SuggestionKind,
+  CompletionItemKind,
+  CompletionItemPriority,
+} from './types';
