+++ src/components/sql-editor/standardSql/standardSuggestionsRegistry.ts
@@ -1,13 +1,29 @@
-// Standard suggestions registry, should work for a standard SQL query.
-// Does not include all functions, only the most common ones.
-
 import { Registry } from '@grafana/data';
-import { TRIGGER_SUGGEST } from '../utils/misc';
-import { CompletionItemPriority, OperatorType, SuggestionKind } from '../utils/types';
-import { AS, ASC, BY, DESC, FROM, GROUP, LIMIT, ORDER, SELECT, STD_STATS, WHERE, WITH } from './language';
+import { TRIGGER_SUGGEST } from '../utils/commands';
+import { CompletionItemPriority, OperatorType, SuggestionKind } from '../types';
+import {
+  AS,
+  ASC,
+  BY,
+  DESC,
+  FROM,
+  GROUP,
+  LIMIT,
+  LOGICAL_OPERATORS,
+  ORDER,
+  SELECT,
+  STD_OPERATORS,
+  STD_STATS,
+  WHERE,
+  WITH,
+} from './language';
 import { FunctionsRegistryItem, OperatorsRegistryItem, SuggestionsRegistyItem } from './types';
 
-// Consumer of the SQL editor should extend this registry with their own custom suggestions.
+/**
+ * This registry glues particular SuggestionKind with an async function that provides completion items for it.
+ * To add a new suggestion kind, SQLEditor should be configured with a provider that implements customSuggestionKinds.
+ */
+
 export const initStandardSuggestions =
   (functions: Registry<FunctionsRegistryItem>, operators: Registry<OperatorsRegistryItem>) =>
   (): SuggestionsRegistyItem[] =>
@@ -16,7 +32,6 @@ export const initStandardSuggestions =
         id: SuggestionKind.SelectKeyword,
         name: SuggestionKind.SelectKeyword,
         suggestions: (_, m) =>
-          (console.log(m) as any) ||
           Promise.resolve([
             {
               label: `${SELECT} <column>`,
@@ -209,105 +224,12 @@ export const initFunctionsRegistry = (): FunctionsRegistryItem[] => [
   })),
 ];
 
-// some std operators...?
-export const initOperatorsRegistry = (): OperatorsRegistryItem[] => [];
-
-// Cloudwatch specific below
-//   case SuggestionKind.Metrics:
-//     {
-//       const namespaceToken = getNamespaceToken(currentToken);
-//       if (namespaceToken?.value) {
-//         // if a namespace is specified, only suggest metrics for the namespace
-//         const metrics = await this.datasource.getMetrics(
-//           this.templateSrv.replace(namespaceToken?.value.replace(/\"/g, '')),
-//           this.templateSrv.replace(this.region)
-//         );
-//         metrics.map((m) => addSuggestion(m.value));
-//       } else {
-//         // If no namespace is specified in the query, just list all metrics
-//         const metrics = await this.datasource.getAllMetrics(this.templateSrv.replace(this.region));
-//         uniq(metrics.map((m) => m.metricName)).map((m) => addSuggestion(m, { insertText: m }));
-//       }
-//     }
-//     break;
-
-//   case SuggestionKind.SchemaKeyword:
-//     addSuggestion(SCHEMA, {
-//       sortText: CompletionItemPriority.High,
-//       insertText: `${SCHEMA}($0)`,
-//       insertTextRules: monaco.languages.CompletionItemInsertTextRule.InsertAsSnippet,
-//       command: TRIGGER_SUGGEST,
-//       kind: monaco.languages.CompletionItemKind.Function,
-//     });
-//     break;
-
-//   case SuggestionKind.Namespaces:
-//     const metricNameToken = getMetricNameToken(currentToken);
-//     let namespaces = [];
-//     if (metricNameToken?.value) {
-//       // if a metric is specified, only suggest namespaces that actually have that metric
-//       const metrics = await this.datasource.getAllMetrics(this.region);
-//       const metricName = this.templateSrv.replace(metricNameToken.value);
-//       namespaces = metrics.filter((m) => m.metricName === metricName).map((m) => m.namespace);
-//     } else {
-//       // if no metric is specified, just suggest all namespaces
-//       const ns = await this.datasource.getNamespaces();
-//       namespaces = ns.map((n) => n.value);
-//     }
-//     namespaces.map((n) => addSuggestion(`"${n}"`, { insertText: `"${n}"` }));
-//     break;
-
-//   case SuggestionKind.LabelKeys:
-//     {
-//       const metricNameToken = getMetricNameToken(currentToken);
-//       const namespaceToken = getNamespaceToken(currentToken);
-//       if (namespaceToken?.value) {
-//         let dimensionFilter = {};
-//         let labelKeyTokens;
-//         if (statementPosition === StatementPosition.SchemaFuncExtraArgument) {
-//           labelKeyTokens = namespaceToken?.getNextUntil(TokenType.Parenthesis, [
-//             TokenType.Delimiter,
-//             TokenType.Whitespace,
-//           ]);
-//         } else if (statementPosition === StatementPosition.AfterGroupByKeywords) {
-//           labelKeyTokens = currentToken?.getPreviousUntil(TokenType.Keyword, [
-//             TokenType.Delimiter,
-//             TokenType.Whitespace,
-//           ]);
-//         }
-//         dimensionFilter = (labelKeyTokens || []).reduce((acc, curr) => {
-//           return { ...acc, [curr.value]: null };
-//         }, {});
-//         const keys = await this.datasource.getDimensionKeys(
-//           this.templateSrv.replace(namespaceToken.value.replace(/\"/g, '')),
-//           this.templateSrv.replace(this.region),
-//           dimensionFilter,
-//           metricNameToken?.value ?? ''
-//         );
-//         keys.map((m) => {
-//           const key = /[\s\.-]/.test(m.value) ? `"${m.value}"` : m.value;
-//           addSuggestion(key);
-//         });
-//       }
-//     }
-//     break;
-
-//   case SuggestionKind.LabelValues:
-//     {
-//       const namespaceToken = getNamespaceToken(currentToken);
-//       const metricNameToken = getMetricNameToken(currentToken);
-//       const labelKey = currentToken?.getPreviousNonWhiteSpaceToken()?.getPreviousNonWhiteSpaceToken();
-//       if (namespaceToken?.value && labelKey?.value && metricNameToken?.value) {
-//         const values = await this.datasource.getDimensionValues(
-//           this.templateSrv.replace(this.region),
-//           this.templateSrv.replace(namespaceToken.value.replace(/\"/g, '')),
-//           this.templateSrv.replace(metricNameToken.value),
-//           this.templateSrv.replace(labelKey.value),
-//           {}
-//         );
-//         values.map((o) =>
-//           addSuggestion(`'${o.value}'`, { insertText: `'${o.value}' `, command: TRIGGER_SUGGEST })
-//         );
-//       }
-//     }
-//     break;
+export const initOperatorsRegistry = (): OperatorsRegistryItem[] => [
+  ...STD_OPERATORS.map((o) => ({
+    id: o,
+    name: o,
+    operator: o,
+    type: OperatorType.Comparison,
+  })),
+  ...LOGICAL_OPERATORS.map((o) => ({ id: o, name: o, operator: o, type: OperatorType.Logical })),
+];
