+++ src/components/sql-editor/utils/commands.ts
